pyproject.toml
[project]
name = "oc-softdesk"
version = "0.1.0"
description = ""
authors = [
    {name = "Noam",email = "noamguez0@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "drf-nested-routers (>=0.95.0,<0.96.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
